+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Format: `<description> (<contributor>, <commit>)`
 
 ## Unreleased
 
+## Added
+
+- Find notes with missing backlinks using `zk list --missing-backlink`
+
 ## Fixed
 
 - Release tarballs now output the program version (by @WhyNotHugo, 344b99f)